requirements.txt
# Backend Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyjwt==2.8.0

# Database
firebase-admin==6.4.0
python-dotenv==1.0.0

# AI & Computer Vision
opencv-python>=4.9.0
ultralytics>=8.3.0
torch>=2.2.0
torchvision>=0.17.0

# Speech Recognition & Text-to-Speech
SpeechRecognition==3.10.1
pyttsx3==2.90
gTTS==2.5.0
sounddevice==0.4.6
soundfile==0.12.1
pyaudio==0.2.14

# Data Processing
numpy==1.26.3
pandas==2.2.0
pillow==10.2.0

# HTTP & API
requests==2.31.0
httpx==0.26.0
aiofiles==23.2.1

# Utilities
pydantic==2.5.3
python-dateutil==2.8.2
pytz==2024.1

# CORS & Middleware
fastapi-cors==0.0.6